+++ core/src/main/java/nl/nn/adapterframework/pipes/IteratingPipe.java
@@ -201,7 +201,7 @@ public abstract class IteratingPipe extends MessageSendingPipe {
 		msgTransformerPool = TransformerPool.configureTransformer(getLogPrefix(null), classLoader, getNamespaceDefs(), getXpathExpression(), getStyleSheetName(), getOutputType(), !isOmitXmlDeclaration(), getParameterList(), false);
 		try {
 			if (StringUtils.isNotEmpty(getStopConditionXPathExpression())) {
-				stopConditionTp=new TransformerPool(XmlUtils.createXPathEvaluatorSource(null,getStopConditionXPathExpression(),"xml",false));
+				stopConditionTp=TransformerPool.getInstance(XmlUtils.createXPathEvaluatorSource(null,getStopConditionXPathExpression(),"xml",false));
 			}
 		} catch (TransformerConfigurationException e) {
 			throw new ConfigurationException(e);