+++ RELEASES.md
@@ -17,6 +17,7 @@ Upcoming
 - Make it possible to disable lastModifiedDelta in CleanupOldFilesPipe
 - Fix WebServiceListener.processRequest() not being called by ServiceDispatcher (fix on previous commit 'Refactor http package to use IPipeLineSession instead of Map...')
 - Fix cache not returning valueXPath transformed value in first call
+- Add possibility to cache and share transformers (to save memory)